media: add a generic V4L2 camera HAL
This patch adds a generic camera HAL, based on the Linux V4L2 API. It
allows android to use any V4L2 capture hardware, supported by the system
as a camera device. This is an initial version with limited capabilities
and some hard-coded assumptions. In the future more features should be
implemented and a way to configure platform-specific parameters, like
paths to front- and rare-facing camera device nodes, should be added.

Change-Id: I46fb85194dad5c2800e43ed969994d6322c5a407